requirements.txt
absl-py==2.1.0
ai-edge-torch==0.1.1
albucore==0.0.23
albumentations==2.0.3
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.1.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
chex==0.1.88
clearml==1.17.1
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
Cython==0.29.37
decorator==5.1.1
dm-tree==0.1.9
etils==1.12.0
filelock==3.17.0
flatbuffers==25.1.24
flax==0.10.2
fsspec==2025.2.0
furl==2.1.3
gast==0.6.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.12.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
jax==0.4.34
jaxlib==0.4.34
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==2.15.0
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.12.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
onnx==1.17.0
onnx-tf==1.10.0
onnx2tf==1.26.7
onnx_graphsurgeon==0.5.5
onnxruntime-gpu==1.20.1
onnxscript==0.1.0
onnxsim==0.4.30
opencv-python==4.11.0.86
opt_einsum==3.4.0
optax==0.2.4
optree==0.14.0
orbax-checkpoint==0.4.4
orderedmultidict==1.0.1
packaging==23.2
pandas==2.2.3
pathlib2==2.3.7.post1
pillow==11.1.0
prompt-toolkit==1.0.14
protobuf==4.25.6
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyInquirer==1.0.3
PyJWT==2.9.0
pyparsing==3.2.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==5.4.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scc4onnx==1.0.6
scikit-image==0.25.1
scipy==1.15.1
simsimd==6.2.1
six==1.17.0
sng4onnx==1.0.4
spo4onnx==1.0.5
stringzilla==3.11.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-addons==0.22.0
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.1
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.23.0
tensorflow_decision_forests==1.8.1
tensorflowjs==4.14.0
tensorstore==0.1.45
termcolor==2.5.0
tf-keras==2.15.0
tifffile==2025.1.10
toolz==1.0.0
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchmetrics==1.6.1
torchvision==0.21.0+cu126
tqdm==4.67.1
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.14.1
wurlitzer==3.1.1
zipp==3.21.0
